@@ -17,30 +17,30 @@
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@playwright/test": "^1.44.1",
-		"@sveltejs/adapter-cloudflare": "^4.4.0",
-		"@sveltejs/kit": "^2.5.10",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
+		"@playwright/test": "^1.48.1",
+		"@sveltejs/adapter-cloudflare": "^4.7.3",
+		"@sveltejs/kit": "^2.7.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/cookie": "^0.6.0",
-		"eslint": "^9.4.0",
+		"eslint": "^9.13.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.39.0",
-		"globals": "^15.0.0",
-		"prettier": "~3.3.1",
-		"prettier-plugin-svelte": "^3.2.4",
-		"sass": "^1.77.4",
-		"svelte": "^4.2.18",
-		"svelte-check": "^3.8.0",
-		"svelte-preprocess": "^5.1.4",
-		"tslib": "^2.6.3",
-		"typescript": "^5.4.5",
-		"typescript-eslint": "^8.0.0-alpha.20",
-		"vite": "^5.2.12"
+		"eslint-plugin-svelte": "^2.45.1",
+		"globals": "^15.11.0",
+		"prettier": "~3.3.3",
+		"prettier-plugin-svelte": "^3.2.7",
+		"sass": "^1.80.3",
+		"svelte": "^5.0.3",
+		"svelte-check": "^4.0.5",
+		"svelte-preprocess": "^6.0.3",
+		"tslib": "^2.8.0",
+		"typescript": "^5.6.3",
+		"typescript-eslint": "^8.10.0",
+		"vite": "^5.4.9"
 	},
 	"dependencies": {
-		"date-fns": "^3.6.0",
+		"date-fns": "^4.1.0",
 		"octokit": "^4.0.2",
-		"slug": "9.1.0"
+		"slug": "10.0.0"
 	},
-	"packageManager": "pnpm@9.1.0+sha256.22e36fba7f4880ecf749a5ca128b8435da085ecd49575e7fb9e64d6bf4fad394"
+	"packageManager": "pnpm@9.12.2"
 }
